Move user-facing bits of interp to interp-api
