+++ amm/interp-api/src/main/scala/ammonite/Stubs.scala
