+++ amm/interp-api/src/main/scala/ammonite/interp/api/APIHolder.scala
@@ -1,0 +1,6 @@
+package ammonite.interp.api
+
+class APIHolder[T >: Null <: AnyRef] {
+  var value0: T = null
+  implicit lazy val value = value0
+}
