+++ amm/interp-api/src/main/scala/ammonite/interp/api/AmmoniteExit.scala
@@ -1,0 +1,8 @@
+package ammonite.interp.api
+
+import scala.util.control.ControlThrowable
+
+/**
+  * Thrown to exit the REPL cleanly
+  */
+case class AmmoniteExit(value: Any) extends ControlThrowable
