+++ amm/src/main/scala/ammonite/main/Scripts.scala
@@ -4,7 +4,7 @@ import java.nio.file.NoSuchFileException
 
 import ammonite.main.Router.{ArgSig, EntryPoint}
 
-import ammonite.runtime.Evaluator.AmmoniteExit
+import ammonite.interp.api.AmmoniteExit
 import ammonite.util.Name.backtickWrap
 import ammonite.util.Util.CodeSource
 import ammonite.util.{Name, Res, Util}
