+++ readme/Repl.scalatex
@@ -545,7 +545,7 @@
           functionality available under the @code{interp} object, which is also
           available in scripts:
 
-        @hl.ref('amm/'interp/'src/'main/'scala/'ammonite/'interp/"InterpAPI.scala", "trait InterpAPI")
+        @hl.ref('amm/"interp-api"/'src/'main/'scala/'ammonite/'interp/'api/"InterpAPI.scala", "trait InterpAPI")
 
       @sect{Utilities}
         @p
