+++ shell/src/main/resources/ammonite/shell/example-predef-bare.sc
@@ -1,6 +1,6 @@
 // Basically a copy of example-predef.sc that's used for the test suite
 import ammonite.ops._
-import ammonite.runtime.tools.IvyThing.scalaBinaryVersion
+import ammonite.interp.api.IvyConstructor.scalaBinaryVersion
 val scalaVersion = scala.util.Properties.versionNumberString
 val ammVersion = ammonite.Constants.version
 interp.load.cp(Path(sys.env("AMMONITE_SHELL"), pwd))
